requirements.txt
livekit-agents
python-dotenv
livekit-api
livekit-agents[google]~=1.2
livekit-plugins-noise-cancellation